Remove version number from debug engine link file
When upgrading the extension, it could happen that we end up with two different link files, one for the old and one for the newer version.
Now, with the version removed from the file name, the new version should always overwrite the link file to point to the new versions binaries. I think that this is always the desired behavior, since I cannot see two separate installations of this extension ever working when loaded at the same time.