Fix communication with debug engine
Establish a sequential order in which we send messages to the new childs debug engine and the parent session.
This makes sure that all messages have been delivered to the child's debug engine before we resume the child and the parent process.
Otherwise we could end up in a state where the child or parent process has already terminated before we had a chance to send all messages to it.