Generalize function call argument extraction
This adds generalized code to read/write function arguments from/to registers and the stack, depending on the calling conventions for the given architecture.

The supports only very limited function signatures (e.g. no floating point or structure parameters) but this is enough for the create process functions we need to support here.