Improve dependencies
Explicitly resolve the transitive dependency to the NETFXSDK of the VS Debug Engine package. Without this, an IDE that relies on the compile_commands.json might complain about a missing `cor.h` header. This is usually not a problem for compilation in the terminal, since the msvc environment sets all kinds of environment variables that hide this error.