Fix use-after-free bug with DataItems
Incorrect construction of ComObjects for the CreateIn/OutInfo data-items lead to use-after-free bugs that made the debugger crash.
This bug was not triggered in Debug mode (for some unknown reason, it is UB after all) which is why it did not fail in CI, but then lead to problems in the released binaries.

Similar to 067fe9c5.